+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_banned_imports/avoid_banned_imports_rule_test.dart
@@ -46,8 +46,8 @@ void main() {
           "import 'package:my_app/ban_folder/something.dart';",
         ],
         messages: [
-          'Avoid banned imports (sample message)',
-          'Avoid banned imports (sample message)',
+          'Avoid banned imports (sample message).',
+          'Avoid banned imports (sample message).',
         ],
       );
     });